Update README.md
# Tim-Hortons
Join My Discord Server For Support and Updates : https://discord.gg/tZDGSNpNxD


**What this is?**
Recently I have gotten into making MLO's. This is my first actually working MLO. I created this for my server Canada Roleplay. In no way is this finished. This is something im working on. I wanted to publish my copy though so I could hear more ideas. Feel free to use this on your server, its optimized and should be fairly lag free. However expect a few bugs, such as I need to add window frames which is something im working on. This MLO is still very cool though, it has a one room inside, with the counter tables and chairs. This MLO was made to look as similar to a Tim Hortons as possable.